Remove switch case

+++ packages/react-scripts/bin/react-scripts.js
@@ -18,48 +18,42 @@ process.on('unhandledRejection', err => {
 const spawn = require('react-dev-utils/crossSpawn');
 const args = process.argv.slice(2);
 
+const scripts = ['build', 'eject', 'start', 'test'];
 const scriptIndex = args.findIndex(
   x => x === 'build' || x === 'eject' || x === 'start' || x === 'test'
 );
 const script = scriptIndex === -1 ? args[0] : args[scriptIndex];
 const nodeArgs = scriptIndex > 0 ? args.slice(0, scriptIndex) : [];
 
-switch (script) {
-  case 'build':
-  case 'eject':
-  case 'start':
-  case 'test': {
-    const result = spawn.sync(
-      'node',
-      nodeArgs
-        .concat(require.resolve('../scripts/' + script))
-        .concat(args.slice(scriptIndex + 1)),
-      { stdio: 'inherit' }
-    );
-    if (result.signal) {
-      if (result.signal === 'SIGKILL') {
-        console.log(
-          'The build failed because the process exited too early. ' +
-            'This probably means the system ran out of memory or someone called ' +
-            '`kill -9` on the process.'
-        );
-      } else if (result.signal === 'SIGTERM') {
-        console.log(
-          'The build failed because the process exited too early. ' +
-            'Someone might have called `kill` or `killall`, or the system could ' +
-            'be shutting down.'
-        );
-      }
-      process.exit(1);
+if (scripts.includes(script)) {
+  const result = spawn.sync(
+    'node',
+    nodeArgs
+      .concat(require.resolve('../scripts/' + script))
+      .concat(args.slice(scriptIndex + 1)),
+    { stdio: 'inherit' }
+  );
+  if (result.signal) {
+    if (result.signal === 'SIGKILL') {
+      console.log(
+        'The build failed because the process exited too early. ' +
+          'This probably means the system ran out of memory or someone called ' +
+          '`kill -9` on the process.'
+      );
+    } else if (result.signal === 'SIGTERM') {
+      console.log(
+        'The build failed because the process exited too early. ' +
+          'Someone might have called `kill` or `killall`, or the system could ' +
+          'be shutting down.'
+      );
     }
-    process.exit(result.status);
-    break;
+    process.exit(1);
   }
-  default:
-    console.log('Unknown script "' + script + '".');
-    console.log('Perhaps you need to update react-scripts?');
-    console.log(
-      'See: https://facebook.github.io/create-react-app/docs/updating-to-new-releases'
-    );
-    break;
+  process.exit(result.status);
+} else {
+  console.log('Unknown script "' + script + '".');
+  console.log('Perhaps you need to update react-scripts?');
+  console.log(
+    'See: https://facebook.github.io/create-react-app/docs/updating-to-new-releases'
+  );
 }
